changed dirs/naming and updated from POEM081
